+++ airbrake/handler.go
@@ -25,8 +25,6 @@ If you want to Flush notices you can do it as you would with Gobrake's notifier 
 package airbrake
 
 import (
-	"net/http"
-
 	"github.com/airbrake/gobrake"
 	"github.com/goph/emperror"
 	"github.com/goph/emperror/internal"
@@ -47,26 +45,11 @@ func NewHandler(projectID int64, projectKey string) *Handler {
 
 // Handle calls the underlying Airbrake notifier.
 func (h *Handler) Handle(err error) {
-	var req *http.Request
+	// Get HTTP request (if any)
+	req, _ := emperror.HttpRequest(err)
 
-	if err, ok := err.(emperror.HttpError); ok {
-		req = err.HttpRequest()
-	} else if err, ok := err.(emperror.Causer); ok { // Look up in the error stack if there is an HTTP error
-		for err != nil {
-			cause := err.Cause()
-
-			if httpErr, ok := cause.(emperror.HttpError); ok {
-				req = httpErr.HttpRequest()
-
-				break
-			}
-
-			err, ok = cause.(emperror.Causer)
-			if !ok {
-				break
-			}
-		}
-	}
+	// Expose the stackTracer interface on the outer error (if there is stack trace in the error)
+	err = emperror.ExposeStackTrace(err)
 
 	notice := h.Notifier.Notice(err, req, 1)