+++ context_test.go
@@ -3,37 +3,32 @@ package emperror_test
 import (
 	"testing"
 
-	"errors"
-
-	. "github.com/goph/emperror"
+	"github.com/goph/emperror"
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestContext(t *testing.T) {
+func TestWith(t *testing.T) {
 	err := errors.New("error")
 
 	kvs := []interface{}{"a", 123}
-	err = With(err, kvs...)
+	err = emperror.With(err, kvs...)
 	kvs[1] = 0 // With should copy its key values
 
-	require.Implements(t, (*Contextor)(nil), err)
-
-	ctx := err.(Contextor).Context()
+	ctx := emperror.Context(err)
 
 	assert.Equal(t, "a", ctx[0])
 	assert.Equal(t, 123, ctx[1])
 	assert.EqualError(t, err, "error")
 }
 
-func TestContext_Multi(t *testing.T) {
+func TestWith_Multiple(t *testing.T) {
 	err := errors.New("")
 
-	err = With(With(err, "a", 123), "b", 321)
+	err = emperror.With(emperror.With(err, "a", 123), "b", 321)
 
-	require.Implements(t, (*Contextor)(nil), err)
-
-	ctx := err.(Contextor).Context()
+	ctx := emperror.Context(err)
 
 	assert.Equal(t, "a", ctx[0])
 	assert.Equal(t, 123, ctx[1])
@@ -41,33 +36,45 @@ func TestContext_Multi(t *testing.T) {
 	assert.Equal(t, 321, ctx[3])
 }
 
-func TestContext_MultiPrefix(t *testing.T) {
+func TestContextor_MissingValue(t *testing.T) {
 	err := errors.New("")
 
-	err = WithPrefix(With(err, "a", 123), "b", 321)
+	err = emperror.With(emperror.With(err, "k0"), "k1")
 
-	require.Implements(t, (*Contextor)(nil), err)
-
-	ctx := err.(Contextor).Context()
-
-	assert.Equal(t, "a", ctx[2])
-	assert.Equal(t, 123, ctx[3])
-	assert.Equal(t, "b", ctx[0])
-	assert.Equal(t, 321, ctx[1])
-}
-
-func TestContext_MissingValue(t *testing.T) {
-	err := errors.New("")
-
-	err = WithPrefix(With(err, "k0"), "k1")
-
-	require.Implements(t, (*Contextor)(nil), err)
-
-	ctx := err.(Contextor).Context()
+	ctx := emperror.Context(err)
 
 	require.Len(t, ctx, 4)
 
 	for i := 1; i < 4; i += 2 {
 		assert.Nil(t, ctx[i])
 	}
+}
+
+func TestContext(t *testing.T) {
+	err := emperror.With(
+		errors.WithMessage(
+			emperror.With(
+				errors.Wrap(
+					emperror.With(
+						errors.New("error"),
+						"key", "value",
+					),
+					"wrapped error",
+				),
+				"key2", "value2",
+			),
+			"another wrapped error",
+		),
+		"key3", "value3",
+	)
+
+	expected := []interface{}{
+		"key", "value",
+		"key2", "value2",
+		"key3", "value3",
+	}
+
+	actual := emperror.Context(err)
+
+	assert.Equal(t, expected, actual)
 }