+++ handler.go
@@ -1,0 +1,11 @@
+package emperror
+
+//go:generate sh -c "CGO_ENABLED=0 go run vendor/github.com/vektra/mockery/cmd/mockery/mockery.go -name=Handler -output . -outpkg emperror_test -testonly -case underscore"
+
+// Handler is responsible for handling an error.
+//
+// This interface allows libraries to decouple from logging and error handling solutions.
+type Handler interface {
+	// Handle takes care of unhandled errors.
+	Handle(err error)
+}