+++ utils_test.go
@@ -4,35 +4,36 @@ import (
 	"testing"
 
 	"errors"
-	. "github.com/goph/emperror"
+
+	"github.com/goph/emperror"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestHandleRecovery(t *testing.T) {
-	handler := new(TestHandler)
+	handler := new(emperror.TestHandler)
 	err := errors.New("error")
 
 	defer func() {
 		assert.EqualError(t, handler.Last(), "error")
 	}()
-	defer HandleRecover(handler)
+	defer emperror.HandleRecover(handler)
 
 	panic(err)
 }
 
 func TestHandleIfErr(t *testing.T) {
-	handler := new(TestHandler)
+	handler := new(emperror.TestHandler)
 	err := errors.New("error")
 
-	HandleIfErr(handler, err)
+	emperror.HandleIfErr(handler, err)
 
 	assert.Equal(t, err, handler.Last())
 }
 
 func TestHandleIfErr_Nil(t *testing.T) {
-	handler := new(TestHandler)
+	handler := new(emperror.TestHandler)
 
-	HandleIfErr(handler, nil)
+	emperror.HandleIfErr(handler, nil)
 
 	assert.NoError(t, handler.Last())
 }