+++ CHANGELOG.md
@@ -7,6 +7,29 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+
+- `ForEachCause` function to be able to loop through all errors in a chain
+- `Context` function to get the context from an error (and all parent errors)
+- `StackTrace` and `ExposeStackTrace` for working with stack trace
+
+### Changed
+
+- **airbrake:** Use `ForEachCause` to find an HTTP request embedded into an error
+- HTTP Request handling
+
+### Fixed
+
+- **airbrake:** Make sure the stack trace is available from the topmost error
+
+### Removed
+
+- `Causer` interface
+- `WithPrefix` didn't really make sense with the decorator pattern
+
+
 ## [0.7.1] - 2018-04-27
 
 ### Changed